+++ package.json
@@ -839,21 +839,19 @@
   "scripts": {
     "verify": "node ./out/build/verify-tools.js",
     "vscode:prepublish": "npm run build",
-    "compile": "tsc -p ./ && npm run build-preview && npm run build:webview",
+    "compile": "tsc -p ./ && npm run build:webview",
     "watch": "tsc -watch -p ./",
-    "clean": "rimraf out && rimraf preview",
+    "clean": "rimraf out",
     "test": "npm run clean && npm run compile && npm run verify && node ./out/build/install-vscode.js && node ./out/build/unit-tests.js",
     "update-deps": "node_modules/.bin/ncu --upgrade --loglevel verbose --packageFile package.json && npm update",
     "coverage:upload": "codecov -f coverage/coverage-final.json",
-    "build": "npm run clean && npm run lint && npm run compile && npm run build:webview",
+    "build": "npm run clean && npm run lint && npm run compile",
     "lint": "eslint . --ext .ts",
     "lint:fix": "eslint . --ext .ts --fix",
-    "build-preview": "webpack-cli --mode production",
-    "preview-watch": "webpack-cli -w --mode development",
     "snippets-build": "node ./out/build/build-snippets.js",
     "icons-build": "node ./out/build/build-svg-icons.js",
-    "build:webview": "webpack-cli --mode production --config ./src/webview/pipeline/webpack.config.js",
-    "watch:webview": "webpack-cli -w --mode development --config ./src/webview/pipeline/webpack.config.js"
+    "build:webview": "webpack-cli --mode production",
+    "watch:webview": "webpack-cli -w --mode development"
   },
   "devDependencies": {
     "@types/byline": "^4.2.31",
@@ -886,7 +884,7 @@
     "decache": "^4.5.1",
     "eslint": "^6.7.2",
     "eslint-plugin-header": "^3.0.0",
-    "file-loader": "^6.0.0",
+    "file-loader": "^6.2.0",
     "glob": "^7.1.3",
     "istanbul": "^0.4.5",
     "mocha": "^6.2.2",
@@ -900,13 +898,13 @@
     "source-map-support": "^0.5.16",
     "style-loader": "^1.2.1",
     "tmp": "0.1.0",
-    "ts-loader": "^6.2.1",
+    "ts-loader": "^8.0.7",
     "typescript": "^3.5.2",
     "vscode-codicons": "0.0.4",
     "vscode-test": "^1.3.0",
     "walker": "^1.0.7",
-    "webpack": "^4.41.6",
-    "webpack-cli": "^3.3.11"
+    "webpack": "^5.3.2",
+    "webpack-cli": "^4.1.0"
   },
   "dependencies": {
     "@kubernetes/client-node": "^0.12.2",
