--- preview-src/tsconfig.json
@@ -1,15 +1,0 @@
-{
-	"extends": "../tsconfig.json",
-	"compilerOptions": {
-		"outDir": "./dist/",
-		"jsx": "react",
-		"lib": [
-			"es2018",
-			"DOM",
-			"DOM.Iterable"
-		]
-	},
-	"include": [
-		"./"
-	]
-}
