--- src/webview/pipeline/tsconfig.json
@@ -1,18 +1,0 @@
-{
-	"compilerOptions": {
-		"module": "commonjs",
-		"target": "es6",
-		"outDir": "out",
-		"sourceMap": true,
-		"rootDir": ".",
-		"jsx": "react",
-		"lib": [
-			"es2018",
-			"DOM",
-			"DOM.Iterable"
-		]
-	},
-	"include": [
-		"./app",
-	]
-}
