+++ webview-tsconfig.json
@@ -1,0 +1,18 @@
+{
+  "extends": "./tsconfig.json",
+  "compilerOptions": {
+    "outDir": "./dist/",
+    "jsx": "react",
+    "lib": [
+      "es2018",
+      "DOM",
+      "DOM.Iterable"
+    ]
+  },
+  "include": [
+    "./src"
+  ],
+  "exclude": [
+    "./src/util"
+  ]
+}
